mynd-brain/requirements.txt
# MYND Brain - Local ML Server
# Optimized for Apple Silicon M2

# Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# ML Core - PyTorch with MPS (Metal) support
torch>=2.1.0
torchvision>=0.16.0

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.35.0

# Graph Neural Networks
torch-geometric>=2.4.0

# Voice Transcription (Whisper)
openai-whisper>=20231117
soundfile>=0.12.1

# Image Understanding (CLIP)
open-clip-torch>=2.23.0
pillow>=10.0.0

# Utilities
numpy>=1.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Optional: for faster inference
# onnxruntime>=1.16.0
